+++ protocol/flows/src/v5/ping.rs
@@ -1,4 +1,4 @@
-use crate::{ctx::FlowContext, flow_trait::Flow};
+use crate::{flow_context::FlowContext, flow_trait::Flow};
 use kaspa_core::debug;
 use p2p_lib::{
     common::ProtocolError,
